+++ TencentCloud/Ess/V20201111/Models/CreateDocumentRequest.cs
@@ -55,6 +55,8 @@ namespace TencentCloud.Ess.V20201111.Models
         /// <ul>
         /// <li>支持自动签传递印章，可通过指定自动签控件id，指定印章id来完成</li>
         /// </ul>
+        /// 注：只有在控制台编辑模板时，<font color="red">归属给发起方</font>的填写控件（如下图）才能在创建文档的时候进行内容填充。
+        /// ![image](https://qcloudimg.tencent-cloud.cn/raw/a54a76a58c454593d06d8e9883ecc9b3.png)
         /// </summary>
         [JsonProperty("FormFields")]
         public FormField[] FormFields{ get; set; }