+++ TencentCloud/Essbasic/V20210526/Models/FlowInfo.cs
@@ -54,7 +54,10 @@ namespace TencentCloud.Essbasic.V20210526.Models
         public FlowApproverInfo[] FlowApprovers{ get; set; }
 
         /// <summary>
-        /// 表单K-V对列表
+        /// 发起方角色的填写控件的填充内容。
+        /// 
+        /// 注：只有在控制台编辑模板时，<font color="red">归属给发起方</font>的填写控件（如下图）才能在创建文档的时候进行内容填充。(<font color="red">白名单功能需要联系对接经理开通，否则模板编辑时无法将填写控件分配给发起方</font>)。
+        /// ![image](https://qcloudimg.tencent-cloud.cn/raw/b1d3978140ee2b44e2c9fdc96e467a5d.png)
         /// </summary>
         [JsonProperty("FormFields")]
         public FormField[] FormFields{ get; set; }