+++ TencentCloud/Rkp/V20191209/RkpClient.cs
@@ -53,6 +53,8 @@ namespace TencentCloud.Rkp.V20191209
         }
 
         /// <summary>
+        /// 产品侧确认风险探针已停售，无收入，并且已经停服。目前服务使用自建redis，不符合规范需要整改下线。
+        /// 
         /// 根据DevicceToken查询OpenID。
         /// </summary>
         /// <param name="req"><see cref="GetOpenIdRequest"/></param>
@@ -63,6 +65,8 @@ namespace TencentCloud.Rkp.V20191209
         }
 
         /// <summary>
+        /// 产品侧确认风险探针已停售，无收入，并且已经停服。目前服务使用自建redis，不符合规范需要整改下线。
+        /// 
         /// 根据DevicceToken查询OpenID。
         /// </summary>
         /// <param name="req"><see cref="GetOpenIdRequest"/></param>
@@ -74,6 +78,8 @@ namespace TencentCloud.Rkp.V20191209
         }
 
         /// <summary>
+        /// 产品侧确认风险探针已停售，无收入，并且已经停服。目前服务使用自建redis，不符合规范需要整改下线。
+        /// 
         /// 获取token接口。
         /// </summary>
         /// <param name="req"><see cref="GetTokenRequest"/></param>
@@ -84,6 +90,8 @@ namespace TencentCloud.Rkp.V20191209
         }
 
         /// <summary>
+        /// 产品侧确认风险探针已停售，无收入，并且已经停服。目前服务使用自建redis，不符合规范需要整改下线。
+        /// 
         /// 获取token接口。
         /// </summary>
         /// <param name="req"><see cref="GetTokenRequest"/></param>
@@ -95,6 +103,8 @@ namespace TencentCloud.Rkp.V20191209
         }
 
         /// <summary>
+        /// 产品侧确认风险探针已停售，无收入，并且已经停服。目前服务使用自建redis，不符合规范需要整改下线。
+        /// 
         /// 腾讯天御设备风险查询接口，输入由客户应用自主采集的设备信息， 通过腾讯大数据风控能力，可以准确根据输入设备信息，还原设备库中的设备ID，并且识别设备的风险，解决客户业务过程中的设备风险，降低企业损失。
         /// </summary>
         /// <param name="req"><see cref="QueryDevAndRiskRequest"/></param>
@@ -105,6 +115,8 @@ namespace TencentCloud.Rkp.V20191209
         }
 
         /// <summary>
+        /// 产品侧确认风险探针已停售，无收入，并且已经停服。目前服务使用自建redis，不符合规范需要整改下线。
+        /// 
         /// 腾讯天御设备风险查询接口，输入由客户应用自主采集的设备信息， 通过腾讯大数据风控能力，可以准确根据输入设备信息，还原设备库中的设备ID，并且识别设备的风险，解决客户业务过程中的设备风险，降低企业损失。
         /// </summary>
         /// <param name="req"><see cref="QueryDevAndRiskRequest"/></param>