+++ TencentCloud/Vod/V20180717/VodClient.cs
@@ -343,6 +343,7 @@ namespace TencentCloud.Vod.V20180717
         }
 
         /// <summary>
+        /// 该 API 已经<font color=red>不再维护</font>，新版 [音画质重生](https://cloud.tencent.com/document/product/266/102571) 接口使用预置模板，详情请参见 [音画质重生模板](https://cloud.tencent.com/document/product/266/102586#50604b3f-0286-4a10-a3f7-18218116aff7)。
         /// 创建音画质重生模板。
         /// </summary>
         /// <param name="req"><see cref="CreateEnhanceMediaTemplateRequest"/></param>
@@ -353,6 +354,7 @@ namespace TencentCloud.Vod.V20180717
         }
 
         /// <summary>
+        /// 该 API 已经<font color=red>不再维护</font>，新版 [音画质重生](https://cloud.tencent.com/document/product/266/102571) 接口使用预置模板，详情请参见 [音画质重生模板](https://cloud.tencent.com/document/product/266/102586#50604b3f-0286-4a10-a3f7-18218116aff7)。
         /// 创建音画质重生模板。
         /// </summary>
         /// <param name="req"><see cref="CreateEnhanceMediaTemplateRequest"/></param>
@@ -511,6 +513,7 @@ namespace TencentCloud.Vod.V20180717
         }
 
         /// <summary>
+        /// 该 API 已经<font color=red>不再维护</font>，新版 [音画质重生](https://cloud.tencent.com/document/product/266/102571) 接口使用预置模板，详情请参见 [音画质重生模板](https://cloud.tencent.com/document/product/266/102586#50604b3f-0286-4a10-a3f7-18218116aff7)。
         /// 创建视频重生模板。
         /// </summary>
         /// <param name="req"><see cref="CreateRebuildMediaTemplateRequest"/></param>
@@ -521,6 +524,7 @@ namespace TencentCloud.Vod.V20180717
         }
 
         /// <summary>
+        /// 该 API 已经<font color=red>不再维护</font>，新版 [音画质重生](https://cloud.tencent.com/document/product/266/102571) 接口使用预置模板，详情请参见 [音画质重生模板](https://cloud.tencent.com/document/product/266/102586#50604b3f-0286-4a10-a3f7-18218116aff7)。
         /// 创建视频重生模板。
         /// </summary>
         /// <param name="req"><see cref="CreateRebuildMediaTemplateRequest"/></param>
@@ -911,6 +915,7 @@ namespace TencentCloud.Vod.V20180717
         }
 
         /// <summary>
+        /// 该 API 已经<font color=red>不再维护</font>，新版 [音画质重生](https://cloud.tencent.com/document/product/266/102571) 接口使用预置模板，详情请参见 [音画质重生模板](https://cloud.tencent.com/document/product/266/102586#50604b3f-0286-4a10-a3f7-18218116aff7)。
         /// 删除音画质重生模板。
         /// </summary>
         /// <param name="req"><see cref="DeleteEnhanceMediaTemplateRequest"/></param>
@@ -921,6 +926,7 @@ namespace TencentCloud.Vod.V20180717
         }
 
         /// <summary>
+        /// 该 API 已经<font color=red>不再维护</font>，新版 [音画质重生](https://cloud.tencent.com/document/product/266/102571) 接口使用预置模板，详情请参见 [音画质重生模板](https://cloud.tencent.com/document/product/266/102586#50604b3f-0286-4a10-a3f7-18218116aff7)。
         /// 删除音画质重生模板。
         /// </summary>
         /// <param name="req"><see cref="DeleteEnhanceMediaTemplateRequest"/></param>
@@ -1104,6 +1110,7 @@ namespace TencentCloud.Vod.V20180717
         }
 
         /// <summary>
+        /// 该 API 已经<font color=red>不再维护</font>，新版 [音画质重生](https://cloud.tencent.com/document/product/266/102571) 接口使用预置模板，详情请参见 [音画质重生模板](https://cloud.tencent.com/document/product/266/102586#50604b3f-0286-4a10-a3f7-18218116aff7)。
         /// 删除视频重生模板。
         /// </summary>
         /// <param name="req"><see cref="DeleteRebuildMediaTemplateRequest"/></param>
@@ -1114,6 +1121,7 @@ namespace TencentCloud.Vod.V20180717
         }
 
         /// <summary>
+        /// 该 API 已经<font color=red>不再维护</font>，新版 [音画质重生](https://cloud.tencent.com/document/product/266/102571) 接口使用预置模板，详情请参见 [音画质重生模板](https://cloud.tencent.com/document/product/266/102586#50604b3f-0286-4a10-a3f7-18218116aff7)。
         /// 删除视频重生模板。
         /// </summary>
         /// <param name="req"><see cref="DeleteRebuildMediaTemplateRequest"/></param>
@@ -1697,6 +1705,7 @@ namespace TencentCloud.Vod.V20180717
         }
 
         /// <summary>
+        /// 该 API 已经<font color=red>不再维护</font>，新版 [音画质重生](https://cloud.tencent.com/document/product/266/102571) 接口使用预置模板，详情请参见 [音画质重生模板](https://cloud.tencent.com/document/product/266/102586#50604b3f-0286-4a10-a3f7-18218116aff7)。
         /// 获取音画质重生模板列表。
         /// </summary>
         /// <param name="req"><see cref="DescribeEnhanceMediaTemplatesRequest"/></param>
@@ -1707,6 +1716,7 @@ namespace TencentCloud.Vod.V20180717
         }
 
         /// <summary>
+        /// 该 API 已经<font color=red>不再维护</font>，新版 [音画质重生](https://cloud.tencent.com/document/product/266/102571) 接口使用预置模板，详情请参见 [音画质重生模板](https://cloud.tencent.com/document/product/266/102586#50604b3f-0286-4a10-a3f7-18218116aff7)。
         /// 获取音画质重生模板列表。
         /// </summary>
         /// <param name="req"><see cref="DescribeEnhanceMediaTemplatesRequest"/></param>
@@ -2114,6 +2124,7 @@ namespace TencentCloud.Vod.V20180717
         }
 
         /// <summary>
+        /// 该 API 已经<font color=red>不再维护</font>，新版 [音画质重生](https://cloud.tencent.com/document/product/266/102571) 接口使用预置模板，详情请参见 [音画质重生模板](https://cloud.tencent.com/document/product/266/102586#50604b3f-0286-4a10-a3f7-18218116aff7)。
         /// 获取视频重生模板列表。
         /// </summary>
         /// <param name="req"><see cref="DescribeRebuildMediaTemplatesRequest"/></param>
@@ -2124,6 +2135,7 @@ namespace TencentCloud.Vod.V20180717
         }
 
         /// <summary>
+        /// 该 API 已经<font color=red>不再维护</font>，新版 [音画质重生](https://cloud.tencent.com/document/product/266/102571) 接口使用预置模板，详情请参见 [音画质重生模板](https://cloud.tencent.com/document/product/266/102586#50604b3f-0286-4a10-a3f7-18218116aff7)。
         /// 获取视频重生模板列表。
         /// </summary>
         /// <param name="req"><see cref="DescribeRebuildMediaTemplatesRequest"/></param>
@@ -2550,6 +2562,7 @@ namespace TencentCloud.Vod.V20180717
         }
 
         /// <summary>
+        /// 该 API 已经<font color=red>不再维护</font>，请使用新版接口 [音画质重生](https://cloud.tencent.com/document/api/266/102571)。
         /// 使用模板发起音画质重生。
         /// </summary>
         /// <param name="req"><see cref="EnhanceMediaByTemplateRequest"/></param>
@@ -2560,6 +2573,7 @@ namespace TencentCloud.Vod.V20180717
         }
 
         /// <summary>
+        /// 该 API 已经<font color=red>不再维护</font>，请使用新版接口 [音画质重生](https://cloud.tencent.com/document/api/266/102571)。
         /// 使用模板发起音画质重生。
         /// </summary>
         /// <param name="req"><see cref="EnhanceMediaByTemplateRequest"/></param>
@@ -2940,6 +2954,7 @@ namespace TencentCloud.Vod.V20180717
         }
 
         /// <summary>
+        /// 该 API 已经<font color=red>不再维护</font>，新版 [音画质重生](https://cloud.tencent.com/document/product/266/102571) 接口使用预置模板，详情请参见 [音画质重生模板](https://cloud.tencent.com/document/product/266/102586#50604b3f-0286-4a10-a3f7-18218116aff7)。
         /// 修改音画质重生模板。
         /// </summary>
         /// <param name="req"><see cref="ModifyEnhanceMediaTemplateRequest"/></param>
@@ -2950,6 +2965,7 @@ namespace TencentCloud.Vod.V20180717
         }
 
         /// <summary>
+        /// 该 API 已经<font color=red>不再维护</font>，新版 [音画质重生](https://cloud.tencent.com/document/product/266/102571) 接口使用预置模板，详情请参见 [音画质重生模板](https://cloud.tencent.com/document/product/266/102586#50604b3f-0286-4a10-a3f7-18218116aff7)。
         /// 修改音画质重生模板。
         /// </summary>
         /// <param name="req"><see cref="ModifyEnhanceMediaTemplateRequest"/></param>
@@ -3163,6 +3179,7 @@ namespace TencentCloud.Vod.V20180717
         }
 
         /// <summary>
+        /// 该 API 已经<font color=red>不再维护</font>，新版 [音画质重生](https://cloud.tencent.com/document/product/266/102571) 接口使用预置模板，详情请参见 [音画质重生模板](https://cloud.tencent.com/document/product/266/102586#50604b3f-0286-4a10-a3f7-18218116aff7)。
         /// 修改视频重生模板。
         /// </summary>
         /// <param name="req"><see cref="ModifyRebuildMediaTemplateRequest"/></param>
@@ -3173,6 +3190,7 @@ namespace TencentCloud.Vod.V20180717
         }
 
         /// <summary>
+        /// 该 API 已经<font color=red>不再维护</font>，新版 [音画质重生](https://cloud.tencent.com/document/product/266/102571) 接口使用预置模板，详情请参见 [音画质重生模板](https://cloud.tencent.com/document/product/266/102586#50604b3f-0286-4a10-a3f7-18218116aff7)。
         /// 修改视频重生模板。
         /// </summary>
         /// <param name="req"><see cref="ModifyRebuildMediaTemplateRequest"/></param>
@@ -3680,6 +3698,7 @@ namespace TencentCloud.Vod.V20180717
         }
 
         /// <summary>
+        /// 该 API 已经<font color=red>不再维护</font>，请使用新版接口 [音画质重生](https://cloud.tencent.com/document/api/266/102571)。
         /// 发起音画质重生
         /// </summary>
         /// <param name="req"><see cref="RebuildMediaRequest"/></param>
@@ -3690,6 +3709,7 @@ namespace TencentCloud.Vod.V20180717
         }
 
         /// <summary>
+        /// 该 API 已经<font color=red>不再维护</font>，请使用新版接口 [音画质重生](https://cloud.tencent.com/document/api/266/102571)。
         /// 发起音画质重生
         /// </summary>
         /// <param name="req"><see cref="RebuildMediaRequest"/></param>
@@ -3701,6 +3721,7 @@ namespace TencentCloud.Vod.V20180717
         }
 
         /// <summary>
+        /// 该 API 已经<font color=red>不再维护</font>，请使用新版接口 [音画质重生](https://cloud.tencent.com/document/api/266/102571)。
         /// 使用模板发起视频重生。
         /// </summary>
         /// <param name="req"><see cref="RebuildMediaByTemplateRequest"/></param>
@@ -3711,6 +3732,7 @@ namespace TencentCloud.Vod.V20180717
         }
 
         /// <summary>
+        /// 该 API 已经<font color=red>不再维护</font>，请使用新版接口 [音画质重生](https://cloud.tencent.com/document/api/266/102571)。
         /// 使用模板发起视频重生。
         /// </summary>
         /// <param name="req"><see cref="RebuildMediaByTemplateRequest"/></param>